@@ -64,7 +64,8 @@
   },
   "dependencies": {
     "ip": "^1.1.4",
-    "react": "^15.2.1",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
     "react-dom": "^15.2.1",
     "react-redux": "^5.0.1",
     "react-router": "^3.0.0",
