@@ -1,5 +1,7 @@
 import React from 'react'
 import Header from '../../components/Header'
+import PropTypes from 'prop-types'
+
 export const CoreLayout = ({ children }) => (
   <div className="container text-center">
     <Header />
@@ -10,7 +12,7 @@ export const CoreLayout = ({ children }) => (
 )
 
 CoreLayout.propTypes = {
-  children: React.PropTypes.element.isRequired,
+  children: PropTypes.element.isRequired,
 }
 
 export default CoreLayout
