@@ -30,7 +30,8 @@ const config = {
     'react',
     'react-redux',
     'react-router',
-    'redux'
+    'redux',
+    'prop-types'
   ]
 }
 
